src/tokenizer/requirements.txt
torch==2.1.0
transformers==4.36.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.3
fastapi==0.104.1
uvicorn==0.24.0
numpy<2.0.0